Product controller test

+++ NLayer.API/Controllers/ProductsController.cs
@@ -60,6 +60,9 @@ namespace NLayer.API.Controllers
         {
             var product = await _service.GetByIdAsync(id);
 
+            if (product == null)
+                return CreateActionResult(CustomResponseDto<NoContentDto>.Fail(404, "Bu id'ye ait ürün bulunamadı"));
+
             await _service.RemoveAsync(product);
 
             return CreateActionResult(CustomResponseDto<NoContentDto>.Success(204));
